@@ -7,7 +7,12 @@ export default abstract class FactList extends JelType {
   constructor() {
     super();
   }
-  
+ /*
+ 	 - the currently valid fact, compiled from assertions and formulars. Ideally a value, but could be something like "greater than 5" if only
+     assertions are available. optional.
+	 - list of Fact, ordered by end-time descending
+
+ */
 
 }
 
